- Removed all errors & warnings when compiling for Win64 (x64)

+++ stb_vorbis/stb_vorbis.c
@@ -4379,7 +4379,7 @@ int stb_vorbis_decode_frame_pushdata(
          while (get8_packet(f) != EOP)
             if (f->eof) break;
          *samples = 0;
-         return f->stream - data;
+         return (int)(f->stream - data);
       }
       if (error == VORBIS_continued_packet_flag_invalid) {
          if (f->previous_length == 0) {
@@ -4389,7 +4389,7 @@ int stb_vorbis_decode_frame_pushdata(
             while (get8_packet(f) != EOP)
                if (f->eof) break;
             *samples = 0;
-            return f->stream - data;
+            return (int)(f->stream - data);
          }
       }
       // if we get an error while parsing, what to do?
@@ -4409,7 +4409,7 @@ int stb_vorbis_decode_frame_pushdata(
    if (channels) *channels = f->channels;
    *samples = len;
    *output = f->outputs;
-   return f->stream - data;
+   return (int)(f->stream - data);
 }
 
 stb_vorbis *stb_vorbis_open_pushdata(
@@ -4432,7 +4432,7 @@ stb_vorbis *stb_vorbis_open_pushdata(
    f = vorbis_alloc(&p);
    if (f) {
       *f = p;
-      *data_used = f->stream - data;
+      *data_used = (int)(f->stream - data);
       *error = 0;
       return f;
    } else {
@@ -4447,7 +4447,7 @@ unsigned int stb_vorbis_get_file_offset(stb_vorbis *f)
    #ifndef STB_VORBIS_NO_PUSHDATA_API
    if (f->push_mode) return 0;
    #endif
-   if (USE_MEMORY(f)) return f->stream - f->stream_start;
+   if (USE_MEMORY(f)) return (int)(f->stream - f->stream_start);
    #ifndef STB_VORBIS_NO_STDIO
    return (unsigned int)orxResource_Tell(f->h) - f->f_start;
    #endif
